requirements.txt
websockets>=11.0.3
aiohttp>=3.8.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.13.0
pydantic>=2.0.0
websocket-client>=1.6.0
requests>=2.31.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
flask>=2.3.0
flask-socketio>=5.3.0
eventlet>=0.33.0
gevent>=23.0.0
gevent-websocket>=0.10.1 